requirements.txt
alembic==1.13.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
contourpy==1.2.0
cryptography==42.0.5
cssselect==1.2.0
cycler==0.12.1
dash==2.16.1
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.0
fastapi-cli==0.0.2
filelock==3.13.4
Flask==3.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fonttools==4.50.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hyperlink==21.0.0
idna==3.6
importlib_metadata==7.0.2
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
joselib==0.1.0b0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
livereload==2.6.3
lxml==5.2.1
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==1.26.4
orjson==3.10.3
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
parsel==1.9.1
passlib==1.7.4
pillow==10.2.0
plotly==5.20.0
Protego==0.3.1
protobuf==4.25.3
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.9.0
PyDispatcher==2.0.7
Pygments==2.17.2
pyOpenSSL==24.1.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
queuelib==1.6.2
referencing==0.35.1
requests==2.31.0
requests-file==2.0.0
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.12.0
Scrapy==2.11.1
selenium==4.20.0
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.28
starlette==0.36.3
streamlit==1.34.0
tenacity==8.2.3
threadpoolctl==3.3.0
tldextract==5.1.2
toml==0.10.2
toolz==0.12.1
tornado==6.4
trio==0.25.0
trio-websocket==0.11.1
Twisted==24.3.0
typer==0.12.3
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
w3lib==2.1.2
watchdog==4.0.0
watchfiles==0.21.0
webdriver-manager==4.0.1
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.18.1
zope.interface==6.3
